# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Scheduler",
    "description": "AI-powered daily news & trading report system",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/retrospective/weekly": {
      "get": {
        "tags": [
          "retrospective"
        ],
        "summary": "List Weekly Analyses",
        "description": "List weekly analyses with pagination.",
        "operationId": "list_weekly_analyses_api_retrospective_weekly_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyAnalysisOut"
                  },
                  "title": "Response List Weekly Analyses Api Retrospective Weekly Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/retrospective/weekly/{analysis_id}": {
      "get": {
        "tags": [
          "retrospective"
        ],
        "summary": "Get Weekly Analysis",
        "description": "Get a specific weekly analysis.",
        "operationId": "get_weekly_analysis_api_retrospective_weekly__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyAnalysisDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/retrospective/daily-checks": {
      "get": {
        "tags": [
          "retrospective"
        ],
        "summary": "List Daily Checks",
        "description": "List recent daily retrospective checks.",
        "operationId": "list_daily_checks_api_retrospective_daily_checks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 14,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyCheckOut"
                  },
                  "title": "Response List Daily Checks Api Retrospective Daily Checks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DailyCheckOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date"
          },
          "recommendations_checked": {
            "type": "integer",
            "title": "Recommendations Checked"
          },
          "targets_hit": {
            "type": "integer",
            "title": "Targets Hit"
          },
          "stops_hit": {
            "type": "integer",
            "title": "Stops Hit"
          },
          "expired_count": {
            "type": "integer",
            "title": "Expired Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "recommendations_checked",
          "targets_hit",
          "stops_hit",
          "expired_count"
        ],
        "title": "DailyCheckOut",
        "description": "Daily retrospective check."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyAnalysisDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "week_start": {
            "type": "string",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "title": "Week End"
          },
          "total_recommendations": {
            "type": "integer",
            "title": "Total Recommendations"
          },
          "win_count": {
            "type": "integer",
            "title": "Win Count"
          },
          "loss_count": {
            "type": "integer",
            "title": "Loss Count"
          },
          "avg_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Return Pct"
          },
          "best_pick_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Pick Ticker"
          },
          "worst_pick_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worst Pick Ticker"
          },
          "analysis_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Text"
          },
          "sector_breakdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Breakdown"
          },
          "lessons": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lessons"
          }
        },
        "type": "object",
        "required": [
          "id",
          "week_start",
          "week_end",
          "total_recommendations",
          "win_count",
          "loss_count",
          "avg_return_pct",
          "best_pick_ticker",
          "worst_pick_ticker",
          "analysis_text",
          "sector_breakdown",
          "lessons"
        ],
        "title": "WeeklyAnalysisDetailOut",
        "description": "Weekly analysis with full detail."
      },
      "WeeklyAnalysisOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "week_start": {
            "type": "string",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "title": "Week End"
          },
          "total_recommendations": {
            "type": "integer",
            "title": "Total Recommendations"
          },
          "win_count": {
            "type": "integer",
            "title": "Win Count"
          },
          "loss_count": {
            "type": "integer",
            "title": "Loss Count"
          },
          "avg_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Return Pct"
          },
          "best_pick_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Pick Ticker"
          },
          "worst_pick_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worst Pick Ticker"
          },
          "analysis_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "week_start",
          "week_end",
          "total_recommendations",
          "win_count",
          "loss_count",
          "avg_return_pct",
          "best_pick_ticker",
          "worst_pick_ticker",
          "analysis_text"
        ],
        "title": "WeeklyAnalysisOut",
        "description": "Weekly analysis summary."
      }
    }
  }
}